Moved methods to Model class and added OWLIndividual support. Renamed test.
