+++ src/java/org/sharegov/cirm/OWL.java
@@ -1344,35 +1344,6 @@ public class OWL
 		return json;
 	}
 	
-	public static String prefixFor(String fragment, Class<?> type) throws Exception
-	{
-		for (String prefix : OWL.prefixManager().getPrefixNames()) {
-			IRI iri = fullIri(prefix + fragment);
-			for (OWLOntology o : OWL.ontologies()) {
-				if (type == OWLClass.class && o.containsClassInSignature(iri)) {
-					return prefix;
-				} else if (type == OWLDataProperty.class
-						&& o.containsDataPropertyInSignature(iri)) {
-					return prefix;
-				} else if (type == OWLObjectProperty.class
-						&& o.containsObjectPropertyInSignature(iri)) {
-					return prefix;
-				}
-			}
-		}
-		throw new Exception("The fragment was not found in any ontology");
-	}
-	
-	
-	public static String prefixFor(String fragment, Class<?> type, String defaultPrefix)
-	{
-		try{
-			return  prefixFor(fragment, type);
-		}catch (Exception e)
-		{
-			return defaultPrefix;
-		}
-	}
 	
 	public static Json prefix(Json json)
 	{
